requirements.txt
# install pytorch, torchvision manually
transformers
datasets
captum
sklearn
numpy
jsonnet
overrides
pytorch-ignite